refactor: move create account modal into accounts workflow
