+++ frontend/src/accounts/AccountsPage.tsx
@@ -2,7 +2,7 @@ import { useMemo, useState } from 'react'
 import { useAuth } from '@/hooks/useAuth'
 import { useAccounts } from '@/api/accounts'
 import { useTaxAdvantagedCategories } from '@/api/taxAdvantagedCategories'
-import CreateAccountModal from '@/components/CreateAccountModal'
+import CreateAccountModal from '@/accounts/components/create-account-modal/CreateAccountModal'
 import Filters from '@/accounts/components/filters/Filters'
 import AccountListSection from '@/accounts/components/AccountListSection'
 import SummaryStatement from '@/accounts/components/summary/SummaryStatement'
